features: Websocket feature to update recycler view list

+++ app/src/main/java/com/example/weatherviewapi/Domains/ResponseData.kt
@@ -3,7 +3,6 @@ package com.example.weatherviewapi.Domains
 import com.example.weatherviewapi.Adapter.DateTypeAdapter
 import com.google.gson.annotations.JsonAdapter
 import com.google.gson.annotations.SerializedName
-import java.time.format.DateTimeFormatter
 import java.util.Date
 
 data class ResponseData (
@@ -11,7 +10,7 @@ data class ResponseData (
     @JsonAdapter(DateTypeAdapter::class)
     val date: Date,
     val hour: String,
-    val humidity: Int,
-    val temperature: Int
+    val humidity: Double,
+    val temperature: Double
 
 )
